Refacto: Split etl file into etl, analysis et open_data
